LG-12039: capture session read from new capture session results for document check results (#10041)
* LG-12039: consider all document related for doc_auth_success?

changelog: Internal, Doc Auth, Doc auth TrueID doc_auth_success? should consider all document business logic.

* LG-12039: minor refactor and test.

* LG-12039: pull in changes from main

* LG-12039: it's questionable to use concern to reduce code complexity. More appropriate to user helper module.

changelog:  Internal, Doc Auth, New doc_auth_success? should include business decision.
